add rocketmq comsumer

+++ src/main/java/com/harmonycloud/repository/AuditRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
  * @author qidong
  * @date 2019/3/15
  */
-public interface AuditRepository extends JpaRepository<CimsAudit, Long> {
+public interface AuditRepository extends JpaRepository<CimsAudit, Integer> {
 }
